Init: cmdline: Disable selinux and add androidboot.hardware to cmdline
androidboot.hardware=universal3470 (copied to ro.hardware by
init) is necessary for the init process to find the fstab file
which is fstab.universal3470. Otherwise the init tries to load
fstab.unknown and will be stuck during the boot process.
Before CM13 the ro.hardware parameter was detected by init by
evaluatin proc/cpuinfo. This is no longer the case.

Change-Id: I81acc49a3f353641f170610bd53ee6566ca9be43